% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{svr_selection}
\alias{svr_selection}
\title{Title}
\usage{
svr_selection(data, n_splits, window_size, gamma_vals, C_vals, epsilon_vals)
}
\arguments{
\item{data}{A vector containing a univariate time series}

\item{epsilon}{}
}
\description{
Title
}
