% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{xgboost_selection}
\alias{xgboost_selection}
\title{Title}
\usage{
xgboost_selection(
  data,
  n_splits,
  window_size,
  eta_vals,
  gamma_vals,
  max_depths,
  lambda_vals
)
}
\arguments{
\item{lambda_vals}{}
}
\description{
Title
}
